Se cambia la carpeta del proyecto
