+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 26
     defaultConfig {
-        applicationId "com.example.silvi.pethouse"
+        applicationId "com.fdi.pad.pethouse"
         minSdkVersion 15
         targetSdkVersion 26
         versionCode 1
