+++ app/src/main/res/layout/activity_reg_contr.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.silvi.pethouse.regContr">
+    tools:context="com.fdi.pad.pethouse.regContr">
 
     <LinearLayout
         android:layout_width="368dp"
